Switch to using a fixed capacity stack.
Benchmarks were showing that we spent a lot of time inside the vm's push
function. This is closer to the book's strategy (although the book uses)
a pointer to the top of the stack instead of a slice.

NOTE: I tried switching callFrame over to a FixedCapacityStack too, but
it didn't appear to be a win. Possibly because we're still checking for
overflow every time we do a call.