Implement a WASM freestanding library.
Biggest challenge here was building an abstraction over stdout and
stderr. Settled on having JS pass in functions for writing to stdout
and stderr as part of the env, and wrapping these in an OutStream. Then
replaced all existing writes to stderr or stdout with writes to this
wrapped outstream.